Warnings cleanup; revert share/block spilitting for backend

+++ src/include/poolcore/backendData.h
@@ -71,7 +71,7 @@ struct shareInfo {
 struct miningRound {
   enum { CurrentRecordVersion = 1 };
   
-  unsigned height;
+  uint64_t height;
   std::string blockHash;
   time_t time;
